backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
motor>=3.1.1
openai>=0.27.0
python-dotenv>=1.0.0
pydantic>=1.10.0